ACE_Framwork/requirements.txt
# ACE Product Intelligence Dependencies

# Core LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Web Framework
streamlit>=1.30.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0

# Utilities
python-dotenv>=1.0.0
